UserTable refactored
