+++ lib/presentation/pages/admin_profile/user_transactions_page.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import '../../widgets/profile_header/profile_header.dart';
-import '../../widgets/user_table.dart';
+import '../../widgets/user_table/user_table.dart';
 
 class UserTransactionsPage extends StatelessWidget {
   final User user;
